# Conflicts:
#	jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/coder/stream/LogFileToAsn1CodecPublisher.java

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetLLB12Builder.java
@@ -8,7 +8,7 @@ public class OffsetLLB12Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetLLB12(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(7).intValue();
+   public static Long offsetLLB12(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(7).longValue();
    }
 }
